Added Tokenizer Extensions for Enhanced Functionality
