+++ sudoklify/src/main/kotlin/dev/teogor/sudoklify/tokenizer/TokenizerExtensions.kt
@@ -1,0 +1,8 @@
+package dev.teogor.sudoklify.tokenizer
+
+internal val Int.tokenizer: Tokenizer
+  get() = if (this < 10) {
+    SingleDigitTokenizer()
+  } else {
+    MultiDigitTokenizer(this)
+  }
